afwefawe

+++ src/benchmarklib/tpcc/procedures/tpcc_delivery.cpp
@@ -13,7 +13,7 @@ TPCCDelivery::TPCCDelivery(const int num_warehouses, BenchmarkSQLExecutor& sql_e
   std::uniform_int_distribution<> carrier_dist{1, 10};
   o_carrier_id = carrier_dist(_random_engine);
 
-  ol_delivery_d = std::time(nullptr);
+  ol_delivery_d = static_cast<int32_t>(std::time(nullptr));
 }
 
 bool TPCCDelivery::execute() {
