+++ src/benchmarklib/tpcc/procedures/tpcc_delivery.hpp
@@ -15,8 +15,8 @@ class TPCCDelivery : public AbstractTPCCProcedure {
 
   // Values generated BEFORE the procedure is executed:
   int32_t w_id;           // Home warehouse ID    [1..num_warehouses]
-  int64_t o_carrier_id;   // Carrier ID           [1..10]
-  int64_t ol_delivery_d;  // Current datetime
+  int32_t o_carrier_id;   // Carrier ID           [1..10]
+  int32_t ol_delivery_d;  // Current datetime
 };
 
 }  // namespace opossum
